weapons spawnage changed

+++ unity_project/test_laby/Assets/Scenes/upgrades/WeaponStats.cs
@@ -6,5 +6,5 @@ public class WeaponStats
   public float damage = 1f;
   public float attackspeed = 1f;
   public float range = 1f;
-  public float rarity = 1f;
+  public int   rarity = 1;
 }
